Update JS experimental renderer to tp3 and allow use of peerDependencies (#225)
* Update web3.js to tp3

* Import the whole web3.js library by default

* Stop exporting shared folder

+++ src/nodes/valueNodes/BytesValueNode.ts
@@ -25,13 +25,13 @@ export function bytesValueNode(
 export function getBytesFromBytesValueNode(node: BytesValueNode): Uint8Array {
   switch (node.encoding) {
     case 'utf8':
-      return getUtf8Encoder().encode(node.data);
+      return getUtf8Encoder().encode(node.data) as Uint8Array;
     case 'base16':
-      return getBase16Encoder().encode(node.data);
+      return getBase16Encoder().encode(node.data) as Uint8Array;
     case 'base58':
-      return getBase58Encoder().encode(node.data);
+      return getBase58Encoder().encode(node.data) as Uint8Array;
     case 'base64':
     default:
-      return getBase64Encoder().encode(node.data);
+      return getBase64Encoder().encode(node.data) as Uint8Array;
   }
 }
